Added a new library interface for LBANN to capture common functions for setting up the model from prototext and restoring from a saved session.  This moves the code from the lbann2.cpp front end model into the LBANN library shared object.
